requirements.txt
Django==4.2.7
python-decouple==3.8
requests==2.31.0
Pillow==10.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
whitenoise==6.5.0
django-environ==0.11.2
dj-database-url==2.1.0
